fixes after 1-st review
